+++ R/get_stations.R
@@ -1,5 +1,4 @@
 
-
 #' Get Stations
 #'
 #' @param run_parallel default: TRUE
@@ -11,60 +10,71 @@
 #' @importFrom tidyr pivot_wider
 #' @importFrom data.table rbindlist
 #' @importFrom rlang .data
+#' @examples
+#' stations <- wasserportal::get_stations()
+#' str(stations)
+#'
+get_stations <- function(run_parallel = TRUE)
+{
+  overview_options <- unlist(get_overview_options())
 
-get_stations <- function(run_parallel = TRUE) {
-
-overview_options <- unlist(get_overview_options())
-
-msg <- sprintf("Importing %d station overviews from Wasserportal Berlin",
-               length(overview_options))
-
-
-if (run_parallel) {
-
-  ncores <- parallel::detectCores() - 1
-
-  cl <- parallel::makeCluster(ncores)
-
-  overview_list <- kwb.utils::catAndRun(
-    messageText = msg,
-    expr = parallel::parLapply(
-      cl, overview_options, function(station_type) {
-        try(wasserportal::get_wasserportal_stations_table(type = station_type))
-      })
+  msg <- sprintf(
+    "Importing %d station overviews from Wasserportal Berlin",
+    length(overview_options)
   )
 
-  parallel::stopCluster(cl)
+  if (run_parallel) {
 
-} else {
+    ncores <- parallel::detectCores() - 1L
 
-  overview_list <-  lapply(overview_options, function(station_type) {
-    try(wasserportal::get_wasserportal_stations_table(type = station_type))
-  })
+    cl <- parallel::makeCluster(ncores)
 
-}
+    overview_list <- kwb.utils::catAndRun(
+      messageText = msg,
+      expr = parallel::parLapply(cl, overview_options, function(type) {
+        try(wasserportal::get_wasserportal_stations_table(type = type))
+      })
+    )
 
+    parallel::stopCluster(cl)
 
+  } else {
 
-overview_df <- data.table::rbindlist(overview_list, fill = TRUE, idcol = "key")
+    overview_list <- lapply(overview_options, function(type) {
+      try(get_wasserportal_stations_table(type = type))
+    })
+  }
 
-metadata <- tidyr::separate(
-  data.frame(key = names(overview_options),
-             station_type = as.vector(overview_options)),
-  .data$key,
-  into = c("water_body", "variable"),
-  sep = "\\.",
-  remove = FALSE)
+  overview_df <- data.table::rbindlist(
+    overview_list,
+    fill = TRUE,
+    idcol = "key"
+  )
 
-overview_df <- dplyr::left_join(overview_df, metadata, by = "key")
+  metadata <- tidyr::separate(
+    data.frame(
+      key = names(overview_options),
+      station_type = as.vector(overview_options)
+    ),
+    .data$key,
+    into = c("water_body", "variable"),
+    sep = "\\.",
+    remove = FALSE
+  )
 
-crosstable <- overview_df %>%
-  dplyr::select("Messstellennummer", "Messstellenname", "station_type") %>%
-  dplyr::mutate(value = "x") %>%
-  tidyr::pivot_wider(names_from = "station_type",
-                     values_from = "value")
+  overview_df <- dplyr::left_join(overview_df, metadata, by = "key")
 
-list(overview_list = overview_list,
-     overview_df = overview_df,
-     crosstable = crosstable)
+  crosstable <- overview_df %>%
+    dplyr::select("Messstellennummer", "Messstellenname", "station_type") %>%
+    dplyr::mutate(value = "x") %>%
+    tidyr::pivot_wider(
+      names_from = "station_type",
+      values_from = "value"
+    )
+
+  list(
+    overview_list = overview_list,
+    overview_df = overview_df,
+    crosstable = crosstable
+  )
 }