+++ R/get_wasserportal_stations_table.R
@@ -7,21 +7,24 @@
 #' @examples
 #' get_overview_options()
 #'
-get_overview_options <- function ()  {
-
-  list(surface_water = list(water_level = "ws",
-                            flow = "df",
-                            temperature = "wt",
-                            conductivity = "lf",
-                            ph = "ph",
-                            oxygen_concentration = "og",
-                            oxygen_saturation = "os"),
-       groundwater = list(level = "gws",
-                          quality = "gwq")
-       )
-
+get_overview_options <- function()
+{
+  list(
+    surface_water = list(
+      water_level = "ws",
+      flow = "df",
+      temperature = "wt",
+      conductivity = "lf",
+      ph = "ph",
+      oxygen_concentration = "og",
+      oxygen_saturation = "os"
+    ),
+    groundwater = list(
+      level = "gws",
+      quality = "gwq"
+    )
+  )
 }
-
 
 #' Wasserportal Berlin: get stations overview table
 #'
@@ -32,9 +35,11 @@ get_overview_options <- function ()  {
 #' @return data frame with master data of selected monitoring stations
 #' @export
 #' @importFrom kwb.utils substSpecialChars
-#' @importFrom rvest html_node html_table
+#' @importFrom rvest html_node html_table html_nodes html_attr
 #' @importFrom stringr str_remove_all
 #' @importFrom xml2 read_html
+#' @importFrom dplyr bind_cols
+#' @importFrom tibble tibble
 #' @examples
 #' types <- wasserportal::get_overview_options()
 #' str(types)
@@ -42,34 +47,58 @@ get_overview_options <- function ()  {
 #' str(sw_l)
 
 get_wasserportal_stations_table <- function (
-  type = get_overview_options()$groundwater$level,
-  url_wasserportal = wasserportal_base_url()
-) {
-
-
-
-
-
+    type = get_overview_options()$groundwater$level,
+    url_wasserportal = wasserportal_base_url()
+)
+{
   if (! is.null(type)) {
     type <- match.arg(type, unlist(get_overview_options()))
   }
 
+  overview_url <- sprintf(
+    "%s/messwerte.php?anzeige=tabelle&thema=%s",
+    url_wasserportal,
+    type
+  )
 
-overview_url <- sprintf("%s/messwerte.php?anzeige=tabelle&thema=%s",
-                        url_wasserportal,
-                        type)
+  html_overview <- xml2::read_html(overview_url)
 
-html_overview <- xml2::read_html(overview_url)
+  overview_table <-  html_overview %>%
+    rvest::html_node(xpath = '//*[@id="pegeltab"]') %>%
+    rvest::html_table()
 
-overview_table <-  html_overview %>%
-  rvest::html_node(xpath = '//*[@id="pegeltab"]') %>%
-  rvest::html_table()
+  stammdaten_link <- html_overview %>%
+    rvest::html_node(xpath = '//*[@id="pegeltab"]') %>%
+    rvest::html_nodes("td") %>%
+    rvest::html_nodes("a") %>%
+    rvest::html_attr("href") %>%
+    stringr::str_extract(
+      pattern = ".*anzeige=i.*|.*pegelonline.*|.*brandenburg.*"
+    )
 
-names(overview_table) <- stringr::str_remove_all(names(overview_table),
-                                                 "-") %>%
-  kwb.utils::substSpecialChars()
+  stammdaten_link <- stammdaten_link[!is.na(stammdaten_link)]
 
-overview_table
+  is_wasserportal <- startsWith(stammdaten_link, "station.php")
 
+  stammdaten_link[is_wasserportal] <- sprintf(
+    "%s/%s",
+    url_wasserportal,
+    stammdaten_link[is_wasserportal]
+  )
+
+  ### hack to remove otherwise duplicated Brandenburg master data in case of
+  ### type = c(2surface_water.water_level" = "ws"), i.e.
+  ### "https://pegelportal.brandenburg.de/messstelle.php?fgid=6&pkz=<messstellennummer>&thema=ws_graph"
+  stammdaten_link <- stammdaten_link[!duplicated(stammdaten_link)]
+
+  names(overview_table) <- names(overview_table) %>%
+    stringr::str_remove_all("-") %>%
+    kwb.utils::substSpecialChars()
+
+  stopifnot(nrow(overview_table) == length(stammdaten_link))
+
+  dplyr::bind_cols(
+    overview_table,
+    tibble::tibble(stammdaten_link = stammdaten_link)
+  )
 }
-