+++ R/read_wasserportal_raw.R
@@ -7,10 +7,12 @@
 #'
 #' @importFrom dplyr select_if
 #'
-get_station_variables <- function(station_df) {
-    station_df <- station_df %>%
-    dplyr::select_if(function(x){!all(is.na(x))})
-  names(station_df)[!names(station_df) %in% c("Messstellennummer", "Messstellenname")]
+get_station_variables <- function(station_df)
+{
+  station_df %>%
+    dplyr::select_if(function(x){!all(is.na(x))}) %>%
+    names() %>%
+    setdiff(c("Messstellennummer", "Messstellenname"))
 }
 
 # read_wasserportal_raw --------------------------------------------------------
@@ -27,51 +29,72 @@ get_station_variables <- function(station_df) {
 #' i.e. `get_stations()$crosstable`
 #' @return ????
 #' @export
-#' @import kwb.utils
+#' @importFrom kwb.utils catAndRun selectColumns selectElements
 #' @importFrom kwb.datetime textToEuropeBerlinPosix
+#' @importFrom httr content POST
 read_wasserportal_raw <- function(
-  variable, station, from_date, type = "single", include_raw_time = FALSE,
+  variable,
+  station,
+  from_date,
+  type = "single",
+  include_raw_time = FALSE,
   handle = NULL,
   stations_crosstable
 )
 {
   #variable <- variables[1]
+
+  stopifnot(length(station) == 1L)
+  stopifnot(length(variable) == 1L)
+
   from_date <- assert_date(from_date)
 
+  station_ids <- kwb.utils::selectColumns(
+    stations_crosstable,
+    "Messstellennummer"
+  )
 
-  stopifnot(length(station) == 1)
-  station_ids <- stations_crosstable$Messstellennummer
   stopifnot(station %in% station_ids)
 
-  stopifnot(length(variable) == 1)
-
-  station_df <- stations_crosstable[stations_crosstable$Messstellennummer == station, ] %>%
+  station_df <- stations_crosstable[station_ids == station, , drop = FALSE] %>%
     dplyr::select_if(function(x){!all(is.na(x))})
 
   variable_ids <- get_station_variables(station_df)
+
   stopifnot(variable %in% variable_ids)
 
   sreihe <- kwb.utils::selectElements(elements = type, list(
-    single = "w", single_all = "wa", daily = "m", monthly = "j"
+    single = "w",
+    single_all = "wa",
+    daily = "m",
+    monthly = "j"
   ))
 
   variable <- kwb.utils::selectElements(elements = variable, list(
-    ws = "w", df = "d", wt = "t", lf = "l", ph = "p", og = "o", os = "s"
+    ws = "w",
+    df = "d",
+    wt = "t",
+    lf = "l",
+    ph = "p",
+    og = "o",
+    os = "s"
   ))
 
-  progress <- get_wasserportal_text(station, variable, station_ids, variable_ids)
-  url <- get_wasserportal_url(station, variable)
-
-  # Format the start date
-  sdatum <- format(from_date, format = "%d.%m.%Y")
-
   # Compose the body of the request
-  body <- list(sreihe = sreihe, smode = "c", sdatum = sdatum)
+  body <- list(
+    sreihe = sreihe,
+    smode = "c",
+    sdatum = format(from_date, format = "%d.%m.%Y") # start date
+  )
 
   # Post the request to the web server
   response <- kwb.utils::catAndRun(
-    progress,
-    httr::POST(url, body = body, handle = handle)
+    get_wasserportal_text(station, variable, station_ids, variable_ids),
+    httr::POST(
+      url = get_wasserportal_url(station, variable),
+      body = body,
+      handle = handle
+    )
   )
 
   if (httr::http_error(response)) {
@@ -83,10 +106,10 @@ read_wasserportal_raw <- function(
   text <- httr::content(response, as = "text", encoding = "Latin1")
 
   # Split the text into separate lines
-  textlines <- strsplit(text, "\n")[[1]]
+  textlines <- strsplit(text, "\n")[[1L]]
 
   # Split the header row into fields
-  header_fields <- as.character(read(textlines[1]))
+  header_fields <- as.character(read(textlines[1L]))
 
   # Return empty list with metadata if no data rows are available
   if (length(textlines) == 1L) {
@@ -94,7 +117,7 @@ read_wasserportal_raw <- function(
   }
 
   # Read the data rows
-  data <- read(text, header = FALSE, skip = 1)
+  data <- read(text, header = FALSE, skip = 1L)
 
   # Get the numbers of the data columns
   if (!type %in% c("daily", "monthly")) {
@@ -114,15 +137,14 @@ read_wasserportal_raw <- function(
 }
 
 # clean_timestamp_columns ------------------------------------------------------
+
 clean_timestamp_columns <- function(data, include_raw_time)
 {
   raw_timestamps <- kwb.utils::selectColumns(data, "Datum")
 
   data <- kwb.utils::renameColumns(data, list(Datum = "timestamp_raw"))
 
-  data$timestamp_corr <- repair_wasserportal_timestamps(
-    timestamps = raw_timestamps
-  )
+  data$timestamp_corr <- repair_wasserportal_timestamps(raw_timestamps)
 
   data <- remove_remaining_duplicates(data)
 
@@ -226,12 +248,12 @@ remove_remaining_duplicates <- function(data)
 }
 
 # remove_timestep_outliers -----------------------------------------------------
-remove_timestep_outliers <- function(data, timestamps, timestep = 15 * 60)
+remove_timestep_outliers <- function(data, timestamps, timestep = 15L * 60L)
 {
   stopifnot(inherits(timestamps, "POSIXct"))
   stopifnot(nrow(data) == length(timestamps))
 
-  is_outlier <- as.numeric(timestamps) %% timestep != 0
+  is_outlier <- as.numeric(timestamps) %% timestep != 0L
 
   if (! any(is_outlier)) {
     return(data)