+++ R/wasserportal_base_url.R
@@ -1,0 +1,8 @@
+#' Helper function: Base Url of Berlin Wassersportal
+#'
+#' @return string with base url of Berlin Wasserportal
+#' @export
+
+wasserportal_base_url <- function() {
+  "https://wasserportal.berlin.de"
+}